Extract PointerService
